% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_cats}
\alias{count_cats}
\title{Counting Categories}
\usage{
count_cats(data, cat_vars)
}
\arguments{
\item{data}{dataset that contains the variables}

\item{cat_vars}{names of categorical variables in a character vector}
}
\value{
number of categories in the dataset
}
\description{
Count the number of categories in a categorical variable, or count
the categories of multiple categorical variables.
}
\examples{
count_cats(ggplot2::diamonds, c("cut"))
count_cats(ggplot2::diamonds, c("cut", "color"))
}
