% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_obs}
\alias{count_obs}
\title{Counting Observations in a Category}
\usage{
count_obs(data, cat_vars, fxn = c(min, max))
}
\arguments{
\item{data}{dataset that contains the variables}

\item{cat_vars}{names of categorical variables in a character vector}

\item{fxn}{the min or max function}
}
\value{
number of observation
}
\description{
Count the number of observations in the smallest/largest category
within a categorical variable, or the smallest/largest category defined by multiple
categorical variables.
}
\examples{
count_obs(ggplot2::diamonds, c("cut"), min)
count_obs(ggplot2::diamonds, c("cut", "color"), max)
}
