% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fun_plots.R
\name{create_canvas}
\alias{create_canvas}
\title{Create a Canvas for a Plot}
\usage{
create_canvas(w = 100, h = 100)
}
\arguments{
\item{w}{width of the canvas}

\item{h}{height of the canvas}
}
\value{
an empty canvas for drawing plots
}
\description{
Create a canvas for a plot with ggplot2
}
