% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fun_plots.R
\name{create_ribbon}
\alias{create_ribbon}
\title{Create a Ribbon Plot}
\usage{
create_ribbon(p, w = 50, h = 30, legend = TRUE)
}
\arguments{
\item{p}{ggplot2 object where the ribbon plot should be added to}

\item{w}{width of the canvas}

\item{h}{height of the canvas}

\item{legend}{boolean indicating whether the legend should be displayed}
}
\value{
a plot with ribbons
}
\description{
Create a fun plot with geom_ribbon from ggplot2
}
