% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_plot}
\alias{label_plot}
\title{Labeling the Plot}
\usage{
label_plot(plot_type, ..., label_fmt = TRUE)
}
\arguments{
\item{plot_type}{the type of plot used to visualize data in a string}

\item{...}{variable names in strings}

\item{label_fmt}{Boolean to indicate whether plot labels should be formatted}
}
\value{
labels for a plot
}
\description{
Label the axses of a plot with formatted/unformatted variable names
and label the plot with a formatted/unformatted plot title.
}
\examples{
label_plot("Scatter Plot", "price", "carat", "cut", label_fmt = TRUE)

}
