% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_awss3.R
\name{load_s3_files}
\alias{load_s3_files}
\title{Load files from an aws s3 path}
\usage{
load_s3_files(s3_path = "s3://my_bucket/directory/", file_format = "csv")
}
\arguments{
\item{s3_path}{aws s3 path as a string}

\item{file_format}{file format as a string, either csv, tsv, or xlsx}
}
\value{
a list of dataframes
}
\description{
Load csv, tsv, or xlsx files stored in an aws s3 path
}
