% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_cat_vars.R
\name{make_barplot1}
\alias{make_barplot1}
\title{Barplot for Categorical Var}
\usage{
make_barplot1(data, cat_var, label_fmt = TRUE)
}
\arguments{
\item{data}{dataset that contains the categorical variable}

\item{cat_var}{name of the categorical variable as a string}

\item{label_fmt}{boolean to indicate whether plot labels should be formatted}
}
\value{
none
}
\description{
Create a barplot for a categorical variable with percent and count
labels. Good for variable that has ten or less categories. Label size is adjusted
based on the number of categories.
}
\examples{
make_barplot1(ggplot2::diamonds, "cut")
}
