% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_cat_vars.R
\name{make_barplot2}
\alias{make_barplot2}
\title{Barplot for Categorical by Categorical}
\usage{
make_barplot2(data, cat_var, grp_var)
}
\arguments{
\item{data}{dataset that contains the categorical variables}

\item{cat_var}{name of a categorical variable to be plotted as a string}

\item{grp_var}{name of a grouping categorical variable as a string}
}
\value{
none
}
\description{
Create a barplot for a categorical variable grouped by another
categorical variable, with percent and count labels. Good for plotting ten or
less total categories.
}
