% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_barplot_cat.R
\name{make_barplot_cat}
\alias{make_barplot_cat}
\title{Create barplot for categorical variable.}
\usage{
make_barplot_cat(data, cat_var, xlab)
}
\arguments{
\item{data}{dataset that contains the categorical variables}

\item{cat_var}{name of a categorical variable to be plotted as a string}

\item{xlab}{a string that labels the x-axis}
}
\value{
non
}
\description{
Create a barplot for a categorical variables with percent and count labels.
}
\examples{
make_barplot_cat(diamonds, "cut", "Cut of The Diamonds")
}
