% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_barplot_cat.R
\name{make_barplot_one}
\alias{make_barplot_one}
\title{Barplot for categorical variable.}
\usage{
make_barplot_one(data, cat_var, xlab)
}
\arguments{
\item{data}{dataset that contains the categorical variable}

\item{cat_var}{name of the categorical variable as a string}

\item{xlab}{x-axis label as a string}
}
\value{
none
}
\description{
Create a barplot for a categorical variable with percent and count
labels. Good for variable that has ten or less categories. Label size is adjusted
based on the number of categories.
}
\examples{
make_barplot_one(ggplot2::diamonds, "cut", "Cut of The Diamonds")
}
