% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_cat_vars.R
\name{make_barplot_two}
\alias{make_barplot_two}
\title{Create Barplot for Categorical Variable by Categorical Variable}
\usage{
make_barplot_two(data, cat_var, grp_var, xlab)
}
\arguments{
\item{data}{dataset that contains the categorical variables}

\item{cat_var}{name of a categorical variable to be plotted as a string}

\item{grp_var}{name of a grouping categorical variable as a string}

\item{xlab}{a string that labels the x-axis}
}
\value{
none
}
\description{
Create a barplot for a categorical variable grouped by another
categorical variable, with percent and count labels. Good for plotting ten or
less total categories.
}
