% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_num_vars.R
\name{make_boxplot2}
\alias{make_boxplot2}
\title{Boxplot for Numeric by Two Categorical Variables}
\usage{
make_boxplot2(data, num_var, cat_var, grp_var, plot_title)
}
\arguments{
\item{data}{dataset that contains the variables}

\item{num_var}{name of the numeric variable as a string}

\item{cat_var}{name of the categorical variable as a string}

\item{grp_var}{name of the categorical variable as a string}

\item{plot_title}{title for the plot as a string}
}
\value{
none
}
\description{
Create a boxplot to compare the distribution of a numeric variable
grouped by a categorical variable among the different levels of another categorical
variable. Good for categorical variables that has ten
or less categories. When the minimum observations for a category is < 200, each
observation will be plotted and outliers will be hidden, vice versa for
observations >= 200. Median for each category is labeled in the box.
}
\examples{
make_boxplot2(ggplot2::diamonds, "price", "cut", "color", "Price of Diamonds For by Cuts and Color")
}
