% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_cat_vars.R
\name{make_hbarplot}
\alias{make_hbarplot}
\title{Horizontal Barplot for Categorical Var}
\usage{
make_hbarplot(data, cat_var, label_fmt = TRUE)
}
\arguments{
\item{data}{dataset that contains the categorical variable}

\item{cat_var}{name of the categorical variable to be plotted as a string}

\item{label_fmt}{boolean to indicate whether plot labels should be formatted}
}
\value{
none
}
\description{
Create a horizontal barplot for a categorical variable with percent
and count labels. Good for variable that has greater than ten categories.
}
\examples{
make_hbarplot(ggplot2::diamonds, "cut")
}
