% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_num_vars.R
\name{make_histogram}
\alias{make_histogram}
\title{Histogram for numeric variable}
\usage{
make_histogram(data, num_var, plot_title)
}
\arguments{
\item{data}{dataset that contains the variables}

\item{num_var}{name of the numeric variable as a string}

\item{plot_title}{title for the plot as a string}
}
\value{
none
}
\description{
Create a histogram to display the distribution of a numeric variable.
}
\examples{
make_histogram(ggplot2::diamonds, "price", "Diamond Price Distribution")
}
