% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_num_vars.R
\name{make_scatter_plot}
\alias{make_scatter_plot}
\title{Scatter Plot for Numeric Variables}
\usage{
make_scatter_plot(data, num_var1, num_var2, label_fmt = TRUE)
}
\arguments{
\item{data}{dataset that contains the variables}

\item{num_var1}{name of the numeric variable as a string}

\item{num_var2}{name of the numeric variable as a string}

\item{label_fmt}{boolean to indicate whether plot labels should be formatted}
}
\value{
none
}
\description{
Create a scatter plot to examine relationship between two numeric
variables.
}
\examples{
make_scatter_plot(ggplot2::diamonds, "price", "carat")
}
