% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_num_vars.R
\name{make_scatter_plot1}
\alias{make_scatter_plot1}
\title{Scatter Plot for Numeric Vars}
\usage{
make_scatter_plot1(data, num_var1, num_var2, trend = "lm", label_fmt = TRUE)
}
\arguments{
\item{data}{dataset that contains the variables}

\item{num_var1}{name of the numeric variable as a string}

\item{num_var2}{name of the numeric variable as a string}

\item{trend}{string represents smoothing method to use}

\item{label_fmt}{Boolean to indicate whether plot labels should be formatted}
}
\value{
none
}
\description{
Create a scatter plot to examine relationship between two numeric
variables.
}
\examples{
make_scatter_plot1(ggplot2::diamonds, "carat", "price")
}
