% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tabs.R
\name{make_tab1}
\alias{make_tab1}
\title{Table One}
\usage{
make_tab1(
  data,
  grp_var,
  all = TRUE,
  select = c(),
  labs = FALSE,
  fmt_labs = c()
)
}
\arguments{
\item{data}{dataset that contains the variables of interest}

\item{grp_var}{name of grouping variable for data summary calculation as a string}

\item{all}{boolean to indicate whether select all variables from the dataset}

\item{select}{list of variables names to be selected from the dataset in strings}

\item{labs}{boolean to indicate whether customized labels are being used in the
table}

\item{fmt_labs}{list of customized labels either in strings or in formula format}
}
\value{
none
}
\description{
Create a nicely formatted table for variables of interest.
}
\examples{
make_tab1(ggplot2::diamonds, "cut")

}
