% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tabs.R
\name{make_xtab}
\alias{make_xtab}
\title{Two by Two for Categorical Var}
\usage{
make_xtab(data, cat_var1, cat_var2)
}
\arguments{
\item{data}{dataset that contains the categorical variables}

\item{cat_var1}{name of the categorical variable as a string}

\item{cat_var2}{name of the categorical variable as a string}
}
\value{
none
}
\description{
Create a nicely formatted 2x2 table for categorical variables.
}
\examples{
make_xtab(ggplot2::diamonds, "cut", "color")
}
