% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{plot_roc}
\alias{plot_roc}
\title{ROC for Model Performance}
\usage{
plot_roc(..., plot_title)
}
\arguments{
\item{...}{ROC objects}

\item{plot_title}{title for the plot as a string}
}
\value{
none
}
\description{
Plot ROC curves of logistic regression models to assess model performance.
Curve is labeled with AUC and 95\% CI.
}
