% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tests.R
\name{print_ttest}
\alias{print_ttest}
\title{Student's T-test Result}
\usage{
print_ttest(test)
}
\arguments{
\item{test}{list of the result from a student's t-test}
}
\value{
none
}
\description{
Tidy the printout of a student's t-test in a R markdown.
}
