% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_num_vars.R
\name{make_histogram}
\alias{make_histogram}
\title{Histogram for Numeric Var}
\usage{
make_histogram(data, num_var, bins = 30)
}
\arguments{
\item{data}{dataset that contains the variables}

\item{num_var}{name of the numeric variable as a string}

\item{bins}{number of bins, default to 30}
}
\value{
none
}
\description{
Create a histogram to display the distribution of a numeric variable.
}
\examples{
make_histogram(ggplot2::diamonds, "price")
}
