% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_num_vars.R
\name{make_scatter_plot2}
\alias{make_scatter_plot2}
\title{Scatter Plot for Numeric Grouped by Categorical}
\usage{
make_scatter_plot2(
  data,
  num_var1,
  num_var2,
  grp_var,
  trend = "lm",
  split = FALSE,
  label_fmt = TRUE
)
}
\arguments{
\item{data}{dataset that contains the variables}

\item{num_var1}{name of the numeric variable as a string}

\item{num_var2}{name of the numeric variable as a string}

\item{grp_var}{name of the categorical variable as a string}

\item{trend}{string represents smoothing method to use}

\item{label_fmt}{Boolean to indicate whether plot labels should be formatted}
}
\value{
none
}
\description{
Create a scatter plot to examine relationship between two numeric
variables within the categories of a categorical variable.
}
\examples{
make_scatter_plot2(ggplot2::diamonds, "carat", "price", "cut")
}
