% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tests.R
\name{print_chisq}
\alias{print_chisq}
\title{Chi-squared Test Result}
\usage{
print_chisq(test)
}
\arguments{
\item{test}{list of the result from a chi-squared test}
}
\value{
none
}
\description{
Tidy the printout of a chi-squared test in a R markdown.
}
